Archive/comfyui-prompt-library-session-backup/pyproject.toml
[project]
name = "comfyui-prompt-library"
version = "1.0.0"
description = "A ComfyUI custom node for managing and browsing prompt history with LoRA tracking"
authors = [{name = "Your Name"}]
requires-python = ">=3.8"
dependencies = [
    "Pillow>=9.0.0",
    "aiohttp>=3.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/comfyui-prompt-library"
Repository = "https://github.com/yourusername/comfyui-prompt-library"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
